1. Refactor to make test results injection simpler 2. Updated documentation with test results api 3. Release v0.67.1
