+++ README.md
@@ -54,5 +54,8 @@ method to run tests.
 `setExpectations(stubPath: string, stubServerBaseUrl?: string): Promise<boolean>` <br />
 method to dynamically set stub expectiona. Stub should be running before invoking this method.
 
+`showTestResults = (testFn: (name: string, cb: () => void) => void)` <br />
+method to report test results in any framework so that it shows up in IDE test results interface.
+
 `printSpecmaticJarVersion()` <br />
 method to print the version of specmatic.jar
