+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "specmatic",
-  "version": "0.67.0",
+  "version": "0.67.1",
   "description": "Node wrapper for Specmatic",
   "main": "dist/index.js",
   "scripts": {
