+++ src/bin/core.ts
